+++ app/interactors/authenticate_user.rb
@@ -10,7 +10,7 @@ class AuthenticateUser
   private
 
   def authenticated_user!
-    context.warden.request.env['devise.skip_trackable'] = false
+    context.warden.request.env["devise.skip_trackable"] = false
     context.warden.authenticate!(OPTIONS)
   end
 end