+++ config/initializers/wrap_parameters.rb
@@ -8,5 +8,5 @@ ActiveSupport.on_load(:action_controller) do
   include ActionController::ParamsWrapper
 
   # Enable parameter wrapping for JSON
-  wrap_parameters(format: [:json, :url_encoded_form])
+  wrap_parameters(format: %i(json url_encoded_form))
 end