+++ lib/rails_api_format/lib/rails_api_format/railtie.rb
@@ -1,10 +1,10 @@
 module RailsApiFormat
   class Railtie < ::Rails::Railtie
-    initializer 'rails_api_format.exceptions_app' do |app|
+    initializer "rails_api_format.exceptions_app" do |app|
       app.config.exceptions_app = RailsPublicExceptions.new(Rails.public_path)
     end
 
-    initializer 'rails_api_format.devise' do
+    initializer "rails_api_format.devise" do
       Devise.setup do |config|
         config.warden do |manager|
           manager.failure_app = DeviseFailureApp
@@ -12,7 +12,7 @@ module RailsApiFormat
       end
     end
 
-    initializer 'rails_api_format.responder' do
+    initializer "rails_api_format.responder" do
       ActionController::API.responder = Responder
     end
   end