+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Unreleased
 
+* Sync Rubocop rules with Rails-Base
 * Replace custom token authentication strategy with [simple_token_authentication](https://github.com/gonzalo-bulnes/simple_token_authentication)
   ([#184](https://github.com/fs/rails-base-api/pull/184))
 * Add id to error response